requirements.txt
anyio==4.9.0
APScheduler==3.11.0
certifi==2025.6.15
charset-normalizer==3.4.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-telegram-bot==22.1
requests==2.32.4
sniffio==1.3.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
supabase==1.0.3